+++ packages/composites/anchor/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @chainlink/anchor-adapter
 
+## 3.4.7
+
+### Patch Changes
+
+- Updated dependencies [afaf0017e]
+  - @chainlink/ea-bootstrap@2.16.0
+  - @chainlink/ea-test-helpers@1.4.0
+  - @chainlink/terra-view-function-adapter@1.3.7
+
 ## 3.4.6
 
 ### Patch Changes