+++ packages/composites/dydx-rewards/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @chainlink/dydx-rewards-adapter
 
+## 1.4.8
+
+### Patch Changes
+
+- Updated dependencies [afaf0017e]
+  - @chainlink/ea-bootstrap@2.16.0
+  - @chainlink/ipfs-adapter@1.3.7
+
 ## 1.4.7
 
 ### Patch Changes