+++ packages/sources/1forge/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @chainlink/1forge-adapter
 
+## 1.6.8
+
+### Patch Changes
+
+- Updated dependencies [afaf0017e]
+  - @chainlink/ea-bootstrap@2.16.0
+  - @chainlink/ea-test-helpers@1.4.0
+
+## 1.6.7
+
+### Patch Changes
+
+- 99a5aad35: Fix validation of batchable property path
+
 ## 1.6.6
 
 ### Patch Changes