+++ packages/sources/dwolla/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @chainlink/dwolla-adapter
 
+## 1.2.7
+
+### Patch Changes
+
+- Updated dependencies [afaf0017e]
+  - @chainlink/ea-bootstrap@2.16.0
+  - @chainlink/ea-test-helpers@1.4.0
+
 ## 1.2.6
 
 ### Patch Changes