+++ packages/sources/lido/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @chainlink/lido-adapter
 
+## 1.0.13
+
+### Patch Changes
+
+- Updated dependencies [afaf0017e]
+  - @chainlink/ea-bootstrap@2.16.0
+  - @chainlink/ea-test-helpers@1.4.0
+
 ## 1.0.12
 
 ### Patch Changes